+++ sandbox/App.tsx
@@ -1,0 +1,3 @@
+// Still haven't found a way to use src/App automatically without this.
+import {App} from './src/App'
+export default App